Add via file support
Arguments may be passed using a via file. The file is specified by
prefixing the file name with a '@' (e.g. '@args.txt'). All arguments in
the file (separated by whitespace) will be inserted into the argument
list in place of the via file. If the file does not exist, the argument
is passed as is.